mongodb post requests are now protected from non-admins

+++ pages/api/random/deletealbum.js
@@ -1,12 +1,18 @@
 import { connectToDatabase } from "../../../lib/mongodb";
+import { getSession } from "next-auth/client";
 
 export default async (req, res) => {
-  const { db } = await connectToDatabase();
-  const {
-    query: { album },
-  } = req;
+  const session = getSession({ req });
+  if (session && session.isAdmin) {
+    const { db } = await connectToDatabase();
+    const {
+      query: { album },
+    } = req;
 
-  await db.collection("albums").deleteOne({ album: album });
+    await db.collection("albums").deleteOne({ album: album });
 
-  res.end("operation attempted");
+    res.end("operation attempted");
+  } else {
+    res.end(401);
+  }
 };
